legacy_streamlit/requirements.txt
streamlit>=1.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
chromadb>=0.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
